// cypress/test/enums.ts
// others
export { E2EAttribute } from '../../src/shared/E2EDataAttributes/enums';

export enum Selector {
  block = 'block',
  blockNode = 'blockNode',
  blockSelected = 'blockSelected',
  componentPanel = 'componentPanel',
  componentPanelBlock = 'componentPanelBlock',
  connector = 'connector',
  connectorLine = 'connectorLine',
  connectorLineSelected = 'connectorLineSelected',
  diagram = 'diagram',
  diagramArea = 'diagramArea',
  moveableElement = 'moveableElement',
  navigationLayerButton = 'navigationLayerButton',
  navigationLayerButtonActive = 'navigationLayerButtonActive',
  strategyResult = 'strategyResult',
}